Updated Index File (Second Upload)
The images were not appearing on the GitHub Website. I went back to the local Index file and made sure that the Path and File name to the image is exactly the same as the Folder and Image name including Case Sensitive meaning upper an lower case letters are exactly the same as file and folder name.